Final Commit Push
Put my name and Information in the CSS and HTML files, fixed the file names. Linked the CSS sheets. Floated the article left and made it have a set width, Made the first line of the article be small capitals and made the T really big and fancy. Set the aside to float left and have a set size. The put all the markers over their spots on the map then made them show information when you hover over them.